feat: v7.0.0 — 3D volumetric lighting, timeline engine, protocol v4
Major release: 3D stage design, spatial effects, bake-to-LED pipeline.

3D Stage (Layout tab):
- Three.js 3D viewport with orbit/zoom/pan, 2D/3D toggle
- Fixtures as grouped objects (sphere + strings + dots move together)
- Projection surfaces: wall, floor, truss, screen (add/resize/color/drag)
- Per-pixel LED positions resolved from fixture geometry
- Live Preview: strings change color to match running show

Fixture Abstraction:
- Children auto-become fixtures on registration
- Rotation override (decouple hardware wiring from stage placement)
- Group fixture type for targeting multiple children as one

Spatial Effects (Actions tab):
- Sphere, plane, box fields with motion paths and easing
- Fixture-local effects wrapping the 14 classic action types
- Blend modes: replace, add, multiply, screen
- Real-time 3D preview in viewport

Timeline Editor (Runtime tab, replaces classic runners):
- Multi-track editor with per-fixture tracks
- Effect clips: drag to position, resize for duration
- Playhead with transport controls
- Bake button with progress modal

Bake Engine:
- Compiles timelines at 40Hz into per-fixture action segments
- Coarse segmentation (max 16 steps) for LoadStepPayload compatibility
- Auto-sync to performers with per-child progress + HTTP verification
- "All Ready" gate before RUNNER_GO

Protocol v4 (breaking change):
- ledStart/ledEnd upgraded from uint8_t to uint16_t
- Supports >255 total LEDs across multi-string ESP32 fixtures
- Sentinel changed from 0xFF to 0xFFFF
- All firmware render functions updated to uint16_t indexing

New files:
- desktop/shared/spatial_engine.py (SpatialResolver + Intersector)
- desktop/shared/bake_engine.py (BinaryBaker + LSQ format)
- docs/USER_MANUAL.md (7-section user manual)
- tests/setup_demo.py (end-to-end demo script)

184 test assertions (up from ~90). Dashboard shows timeline status.

+++ main/Child.cpp
@@ -34,8 +34,8 @@ volatile uint8_t  childActP8b   = 0;
 volatile uint8_t  childActP8c   = 0;
 volatile uint8_t  childActP8d   = 0;
 volatile uint8_t  childActSeq   = 0;
-volatile uint8_t  childActSt[MAX_STR_PER_CHILD];
-volatile uint8_t  childActEn[MAX_STR_PER_CHILD];
+volatile uint16_t childActSt[MAX_STR_PER_CHILD];
+volatile uint16_t childActEn[MAX_STR_PER_CHILD];
 volatile uint8_t  childBrightness = 255;
 
 ChildRunnerStep   childRunner[MAX_CHILD_STEPS];
